readme

+++ src/CosmicVault.sol
@@ -3,8 +3,9 @@
 pragma solidity 0.8.25;
 
 contract CosmicVault {
-    uint vaulting;
-    receive() external payable { 
+    uint256 vaulting;
+
+    receive() external payable {
         vaulting += msg.value;
     }
